Updated this branch to latest /trunk/wvstreams@11695 using this command:
svn merge                                            
svn+ssh://alumnit.ca/svn/{branches/lime/wvstreams@11207,trunk/wvstreams@11695} 

At least, I think I did.  wvvector.h was suspiciously still broken, but
I fixed it by copying it straight from /trunk/wvstreams.

All unit tests pass *except* the dbus ones (probably due to the pre_select()
changes).  I'll fix that next.



git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/uniservice@11696 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ utils/wvhash.cc
@@ -2,7 +2,7 @@
 
 // Note: this hash function is case-insensitive since it ignores the
 // bit in ASCII that defines case.  You may want to take advantage of this.
-unsigned WvHash(const char *s)
+unsigned int WvHash(const char *s)
 {
     unsigned hash = 0, slide, andval;
     if (!s) return 0;
@@ -30,6 +30,6 @@ unsigned WvHash(const int &i)
 
 unsigned WvHash(const void *p)
 {
-    return reinterpret_cast<unsigned>(p);
+    return reinterpret_cast<unsigned long>(p);
 }
